Remove terminal profiles functionality
We don't want user to override qudosh. As of now it isn't possible to
handle different profiles easily.

We will come back to this topic later on when we will make qudosh
able to call custom binary, in that case we will replace actual binaries
with qudosh only.

Some settings were removed as well to prevent unwanted actions.